Use run_environment to set the environment variables
If the ITK recipe is used to create shared libraries, the test_package
needs to locate these shared libraries. On Windows, they can be copied to the bin-directory.
On Linux and macOS the "LD_LIBRARY_PATH", "DYLD_LIBRARY_PATH" and "DYLD_FRAMEWORK_PATH" environment
variables need to be set which is automatically done by using `run_environment=True`.
Copying the libraries is no longer necessary.